+++ packages/bot/src/bot/start/initializer.ts
@@ -20,7 +20,7 @@ import type { CustomClient } from '../../types'
 import { ConfigurationError } from '@lucky/shared/types'
 import { redisClient } from '@lucky/shared/services'
 import { initProviderHealth } from '../../utils/music/search/providerHealth'
-import { musicWatchdogService } from '../../utils/music/watchdog'
+import { musicWatchdogService } from '../../services/musicManagement/watchdog'
 import { birthdayScheduler } from '../../utils/general/birthdayScheduler'
 import { reminderScheduler } from '../../utils/general/reminderScheduler'
 import { supportSessionScheduler } from '../../utils/general/supportSessionScheduler'
