+++ packages/bot/src/functions/music/commands/stop.spec.ts
@@ -29,7 +29,7 @@ jest.mock('../../../utils/music/queueResolver', () => ({
     resolveGuildQueue: (...args: unknown[]) => resolveGuildQueueMock(...args),
 }))
 
-jest.mock('../../../utils/music/watchdog', () => ({
+jest.mock('../../../services/musicManagement/watchdog', () => ({
     musicWatchdogService: {
         markIntentionalStop: (...args: unknown[]) =>
             markIntentionalStopMock(...args),
