+++ packages/bot/src/functions/music/commands/stop.ts
@@ -10,7 +10,7 @@ import {
 } from '../../../utils/command/commandValidations'
 import { resolveGuildQueue } from '../../../utils/music/queueResolver'
 import { createSuccessEmbed } from '../../../utils/general/embeds'
-import { musicWatchdogService } from '../../../utils/music/watchdog'
+import { musicWatchdogService } from '../../../services/musicManagement/watchdog'
 import { musicSessionSnapshotService } from '../../../utils/music/sessionSnapshots'
 import { assertDefined } from '@lucky/shared/utils/guards'
 
@@ -25,7 +25,16 @@ export default new Command({
         const { queue } = resolveGuildQueue(client, interaction.guildId ?? '')
 
         if (!(await requireQueue(queue, interaction))) return
-        if (!(await requireDJRole(interaction, assertDefined(interaction.guildId, 'Guild ID required after requireGuild check')))) return
+        if (
+            !(await requireDJRole(
+                interaction,
+                assertDefined(
+                    interaction.guildId,
+                    'Guild ID required after requireGuild check',
+                ),
+            ))
+        )
+            return
 
         if (queue) {
             musicWatchdogService.markIntentionalStop(queue.guild.id)
