+++ packages/bot/src/handlers/player/lifecycleHandlers.ts
@@ -3,7 +3,7 @@ import type { Client } from 'discord.js'
 import { infoLog, debugLog } from '@lucky/shared/utils'
 import * as voiceStatus from '../../services/VoiceChannelStatusService'
 import { ENVIRONMENT_CONFIG } from '@lucky/shared/config'
-import { musicWatchdogService } from '../../utils/music/watchdog'
+import { musicWatchdogService } from '../../services/musicManagement/watchdog'
 import { musicSessionSnapshotService } from '../../utils/music/sessionSnapshots'
 import { replenishQueue } from '../../utils/music/queueOperations'
 import type { QueueMetadata } from '../../types/QueueMetadata'
