+++ packages/bot/src/handlers/player/queueExhaustion.ts
@@ -1,11 +1,9 @@
 import type { GuildQueue, Track } from 'discord-player'
 import { errorLog } from '@lucky/shared/utils'
-import { musicWatchdogService } from '../../utils/music/watchdog'
+import { musicWatchdogService } from '../../services/musicManagement/watchdog'
 import { musicSessionSnapshotService } from '../../utils/music/sessionSnapshots'
 import * as voiceStatus from '../../services/VoiceChannelStatusService'
-import {
-    setReplenishSuppressed,
-} from '../../utils/music/replenishSuppressionStore'
+import { setReplenishSuppressed } from '../../utils/music/replenishSuppressionStore'
 
 /**
  * Handles queue exhaustion logic shared between finish and skip events.
