--- packages/bot/src/utils/misc/pathUtils.ts
@@ -1,17 +1,0 @@
-import { fileURLToPath } from 'url'
-import { dirname } from 'path'
-
-export function getDirname(importMetaUrl: string): string {
-    return dirname(fileURLToPath(importMetaUrl))
-}
-
-export function getFilename(importMetaUrl: string): string {
-    return fileURLToPath(importMetaUrl)
-}
-
-export function normalizePath(p: string): string {
-    if (process.platform === 'win32' && p.startsWith('/')) {
-        return p.slice(1)
-    }
-    return p
-}
