+++ packages/bot/src/utils/music/idleDisconnect.ts
@@ -1,7 +1,7 @@
 import type { GuildQueue } from 'discord-player'
 import { guildSettingsService } from '@lucky/shared/services'
 import { debugLog, errorLog } from '@lucky/shared/utils'
-import { musicWatchdogService } from './watchdog'
+import { musicWatchdogService } from '../../services/musicManagement/watchdog'
 import { collaborativePlaylistService } from './collaborativePlaylist'
 import type { QueueMetadata } from '../../types/QueueMetadata'
